ghidra-ai-assistant-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
pydantic>=2.9.0
python-dotenv>=1.0.1